better formatting, moved disclaimer component, syntax fixes
